requirements.txt
matplotlib
scikit-learn
plotly
scipy
openai
pinecone-client
pandas